Add Take Operator
