+++ Runtime/Core/React/Operators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44bec8452249a5242a08a905e84ff15b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
