+++ Runtime/Core/React/Operators/Empty.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 582b72bbdc4a3eb45bb8aa2b6a44a82c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
