+++ Runtime/Core/React/Operators/OperatorObservableBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41a95b01210b89b4fab8ad659d797a43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
