+++ Runtime/Core/React/Operators/Take.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff868bbe6bfd7b8438cd6e1584acefed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
