+++ Runtime/Core/React/Operators/WaitFrame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6cdcf591af1fbd429483face16d51c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
